Remove custom abstract main activity (#228)
We have an abstract custom main class which provides us with nothing but confusion. The static part in the code needs to be cut out elsewhere but for now this removes a lot of unnecessary code.
