+++ src/main/java/menion/android/whereyougo/gui/dialog/AboutDialog.java
@@ -10,7 +10,7 @@ import android.webkit.WebView;
 
 import menion.android.whereyougo.R;
 import menion.android.whereyougo.VersionInfo;
-import menion.android.whereyougo.gui.extension.activity.CustomMainActivity;
+import menion.android.whereyougo.gui.activity.MainActivity;
 import menion.android.whereyougo.gui.extension.dialog.CustomDialogFragment;
 import menion.android.whereyougo.preferences.PreferenceValues;
 import menion.android.whereyougo.utils.A;
@@ -21,9 +21,9 @@ public class AboutDialog extends CustomDialogFragment {
     public Dialog createDialog(Bundle savedInstanceState) {
         String buffer = "<div align=\"center\"><h2><b>" + A.getAppName() + "</b></h2></div>" +
                 "<div align=\"center\"><h3><b>" + A.getAppVersion() + "</b></h3></div>" +
-                CustomMainActivity.loadAssetString(PreferenceValues.getLanguageCode() + "_first.html") +
+                MainActivity.loadAssetString(PreferenceValues.getLanguageCode() + "_first.html") +
                 "<br /><br />" +
-                CustomMainActivity.loadAssetString(PreferenceValues.getLanguageCode() + "_about.html") +
+                MainActivity.loadAssetString(PreferenceValues.getLanguageCode() + "_about.html") +
                 "<br /><br />" +
                 VersionInfo.getNews(1, PreferenceValues.getApplicationVersionActual());
         // add info
